XC-220: move tests
