+++ rs/ethereum/cketh/minter/src/eth_logs/mod.rs
@@ -8,6 +8,7 @@ use crate::eth_rpc::{FixedSizeData, Hash};
 use crate::logs::{DEBUG, INFO};
 use crate::numeric::{BlockNumber, Erc20Value, LogIndex, Wei};
 use candid::Principal;
+use hex_literal::hex;
 use ic_canister_log::log;
 use ic_ethereum_types::Address;
 use minicbor::{Decode, Encode};
@@ -18,6 +19,18 @@ pub use parser::{
     Erc20WithSubaccountLogParser, LogParser, ReceivedErc20LogParser, ReceivedEthLogParser,
 };
 pub use scraping::{LogScraping, ReceivedErc20LogScraping, ReceivedEthLogScraping};
+
+// Keccak256("ReceivedEth(address,uint256,bytes32)")
+const RECEIVED_ETH_EVENT_TOPIC: [u8; 32] =
+    hex!("257e057bb61920d8d0ed2cb7b720ac7f9c513cd1110bc9fa543079154f45f435");
+
+// Keccak256("ReceivedErc20(address,address,uint256,bytes32)")
+const RECEIVED_ERC20_EVENT_TOPIC: [u8; 32] =
+    hex!("4d69d0bd4287b7f66c548f90154dc81bc98f65a1b362775df5ae171a2ccd262b");
+
+// Keccak256("ReceivedErc20(address,address,uint256,bytes32,bytes32)")
+const RECEIVED_ERC20_EVENT_WITH_SUBACCOUNT_TOPIC: [u8; 32] =
+    hex!("aef895090c2f5d6e81a70bef80dce496a0558487845aada57822159d5efae5cf");
 
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Decode, Encode)]
 pub struct ReceivedEthEvent {
